chore: consolidate gitlab services
