+++ src/controllers/gitlabAssignedMrController.js
@@ -1,10 +1,10 @@
-const gitlabAssignedMrService = require('../services/gitlabAssignedMrService');
+const gitlabService = require('../services/gitlabService');
 
 async function getAssignedMrs(req, res) {
   const { username } = req.query;
   if (!username) return res.json({ success: false, error: 'Missing username' });
   try {
-    const mrs = await gitlabAssignedMrService.listAssignedMrs(username);
+    const mrs = await gitlabService.listAssignedMrs(username);
     return res.json({ success: true, data: mrs });
   } catch (error) {
     return res.json({ success: false, error: error.message });
