+++ src/controllers/gitlabPipelineController.js
@@ -1,8 +1,8 @@
-const gitlabPipelineService = require('../services/gitlabPipelineService');
+const gitlabService = require('../services/gitlabService');
 
 async function getPipelineStatusTable(req, res) {
   try {
-    const data = await gitlabPipelineService.getPipelineStatusTable();
+    const data = await gitlabService.getPipelineStatusTable();
     return res.json({ success: true, data });
   } catch (e) {
     return res.status(500).json({ success: false, error: e.message });
