+++ src/controllers/gitlabUsersController.js
@@ -1,8 +1,8 @@
-const gitlabUsersService = require('../services/gitlabUsersService');
+const gitlabService = require('../services/gitlabService');
 
 async function getGitlabUsers(req, res) {
   try {
-    const users = await gitlabUsersService.listGitlabUsers();
+    const users = await gitlabService.listGitlabUsers();
     return res.json({ success: true, data: users });
   } catch (error) {
     return res.json({ success: false, error: error.message });
